[CAS-755] StateFlow spike
