+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/ChatDomainImpl.kt
@@ -1,79 +1,42 @@
 package io.getstream.chat.android.livedata
 
 import android.content.Context
-import android.os.Build
-import android.os.Handler
 import androidx.annotation.VisibleForTesting
 import androidx.lifecycle.LiveData
-import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.asLiveData
-import io.getstream.chat.android.client.BuildConfig.STREAM_CHAT_VERSION
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.api.models.FilterObject
-import io.getstream.chat.android.client.api.models.QueryChannelsRequest
 import io.getstream.chat.android.client.api.models.QuerySort
 import io.getstream.chat.android.client.call.Call
-import io.getstream.chat.android.client.call.await
 import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.events.ChatEvent
-import io.getstream.chat.android.client.events.ConnectedEvent
-import io.getstream.chat.android.client.events.MarkAllReadEvent
-import io.getstream.chat.android.client.extensions.enrichWithCid
-import io.getstream.chat.android.client.logger.ChatLogger
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Config
-import io.getstream.chat.android.client.models.Filters.`in`
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.Mute
 import io.getstream.chat.android.client.models.Reaction
 import io.getstream.chat.android.client.models.TypingEvent
 import io.getstream.chat.android.client.models.User
-import io.getstream.chat.android.client.models.UserEntity
 import io.getstream.chat.android.client.parser.StreamGson
-import io.getstream.chat.android.client.socket.SocketListener
 import io.getstream.chat.android.client.utils.Result
-import io.getstream.chat.android.client.utils.SyncStatus
-import io.getstream.chat.android.client.utils.observable.Disposable
 import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
 import io.getstream.chat.android.livedata.controller.ChannelControllerImpl
 import io.getstream.chat.android.livedata.controller.QueryChannelsControllerImpl
-import io.getstream.chat.android.livedata.extensions.applyPagination
-import io.getstream.chat.android.livedata.extensions.isPermanent
-import io.getstream.chat.android.livedata.extensions.users
-import io.getstream.chat.android.livedata.model.ChannelConfig
 import io.getstream.chat.android.livedata.model.SyncState
 import io.getstream.chat.android.livedata.repository.RepositoryFacade
-import io.getstream.chat.android.livedata.repository.builder.RepositoryFacadeBuilder
 import io.getstream.chat.android.livedata.repository.database.ChatDatabase
-import io.getstream.chat.android.livedata.request.AnyChannelPaginationRequest
 import io.getstream.chat.android.livedata.request.QueryChannelPaginationRequest
 import io.getstream.chat.android.livedata.request.QueryChannelsPaginationRequest
-import io.getstream.chat.android.livedata.request.toAnyChannelPaginationRequest
-import io.getstream.chat.android.livedata.service.sync.BackgroundSyncConfig
-import io.getstream.chat.android.livedata.service.sync.SyncProvider
 import io.getstream.chat.android.livedata.usecase.UseCaseHelper
-import io.getstream.chat.android.livedata.utils.DefaultRetryPolicy
 import io.getstream.chat.android.livedata.utils.Event
 import io.getstream.chat.android.livedata.utils.RetryPolicy
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.async
-import kotlinx.coroutines.cancelChildren
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.filterNotNull
-import kotlinx.coroutines.launch
-import java.util.Date
-import java.util.InputMismatchException
-import java.util.UUID
-import java.util.concurrent.ConcurrentHashMap
+import io.getstream.chat.android.offline.ChatDomainImpl as NewChatDomainImpl
 
 private val CHANNEL_CID_REGEX = Regex("^!?[\\w-]+:!?[\\w-]+$")
-private const val MESSAGE_LIMIT = 30
-private const val MEMBER_LIMIT = 30
-private const val INITIAL_CHANNEL_OFFSET = 0
-private const val CHANNEL_LIMIT = 30
 
 internal val gson = StreamGson.gson
 
@@ -96,88 +59,69 @@ internal val gson = StreamGson.gson
  * chatDomain.errorEvents events for errors that happen while interacting with the chat
  *
  */
-internal class ChatDomainImpl internal constructor(
-    internal var client: ChatClient,
-    // the new behaviour for ChatDomain is to follow the ChatClient.setUser
-    // the userOverwrite field is here for backwards compatibility
-    internal var userOverwrite: User? = null,
-    internal var db: ChatDatabase? = null,
-    private val mainHandler: Handler,
-    override var offlineEnabled: Boolean = true,
-    internal var recoveryEnabled: Boolean = true,
-    override var userPresence: Boolean = false,
-    internal var backgroundSyncEnabled: Boolean = false,
-    internal var appContext: Context,
-) :
-    ChatDomain {
-    internal constructor(
-        client: ChatClient,
-        handler: Handler,
-        offlineEnabled: Boolean,
-        recoveryEnabled: Boolean,
-        userPresence: Boolean,
-        backgroundSyncEnabled: Boolean,
-        appContext: Context,
-    ) : this(
-        client,
-        null,
-        null,
-        handler,
-        offlineEnabled,
-        recoveryEnabled,
-        userPresence,
-        backgroundSyncEnabled,
-        appContext
-    )
+internal class ChatDomainImpl internal constructor(private val delegate: NewChatDomainImpl) : ChatDomain {
 
-    private val _initialized = MutableStateFlow(false)
-    private val _online = MutableStateFlow(false)
+    internal var client: ChatClient = delegate.client
 
-    private val _totalUnreadCount = MutableStateFlow(0)
-    private val _channelUnreadCount = MutableStateFlow(0)
-    private val _errorEvent = MutableStateFlow<Event<ChatError>?>(null)
-    private val _banned = MutableStateFlow(false)
-    private val _mutedUsers = MutableStateFlow<List<Mute>>(emptyList())
-    private val _typingChannels = MediatorLiveData<TypingEvent>()
+    internal val appContext: Context
+        get() = delegate.appContext
 
-    override lateinit var currentUser: User
-    lateinit var database: ChatDatabase
-    private val syncModule by lazy { SyncProvider(appContext) }
+    override var offlineEnabled: Boolean
+        get() = delegate.offlineEnabled
+        set(value) {
+            delegate.offlineEnabled = value
+        }
+
+    override var userPresence: Boolean
+        get() = delegate.userPresence
+        set(value) {
+            delegate.userPresence = value
+        }
+
+    override var currentUser: User
+        get() = delegate.currentUser
+        set(value) {
+            delegate.currentUser = value
+        }
+
+    val database: ChatDatabase
+        get() = delegate.database
 
     /** a helper object which lists all the initialized use cases for the chat domain */
     override val useCases: UseCaseHelper = UseCaseHelper(this)
 
     @VisibleForTesting
-    val defaultConfig: Config = Config(isConnectEvents = true, isMutes = true)
+    val defaultConfig: Config
+        get() = delegate.defaultConfig
 
     /** if the client connection has been initialized */
-    override val initialized: LiveData<Boolean> = _initialized.asLiveData()
+    override val initialized: LiveData<Boolean> = delegate.initialized.asLiveData()
 
     /**
      * LiveData<Boolean> that indicates if we are currently online
      */
-    override val online: LiveData<Boolean> = _online.asLiveData()
+    override val online: LiveData<Boolean> = delegate.online.asLiveData()
 
     /**
      * The total unread message count for the current user.
      * Depending on your app you'll want to show this or the channelUnreadCount
      */
-    override val totalUnreadCount: LiveData<Int> = _totalUnreadCount.asLiveData()
+    override val totalUnreadCount: LiveData<Int> = delegate.totalUnreadCount.asLiveData()
 
     /**
      * the number of unread channels for the current user
      */
-    override val channelUnreadCount: LiveData<Int> = _channelUnreadCount.asLiveData()
+    override val channelUnreadCount: LiveData<Int> = delegate.channelUnreadCount.asLiveData()
 
     /**
      * list of users that you've muted
      */
-    override val muted: LiveData<List<Mute>> = _mutedUsers.asLiveData()
+    override val muted: LiveData<List<Mute>> = delegate.muted.asLiveData()
 
     /**
      * if the current user is banned or not
      */
-    override val banned: LiveData<Boolean> = _banned.asLiveData()
+    override val banned: LiveData<Boolean> = delegate.banned.asLiveData()
 
     /**
      * The error event livedata object is triggered when errors in the underlying components occur.
@@ -188,310 +132,55 @@ internal class ChatDomainImpl internal constructor(
      *   })
      *
      */
-    override val errorEvents: LiveData<Event<ChatError>> = _errorEvent.filterNotNull().asLiveData()
+    override val errorEvents: LiveData<Event<ChatError>> = delegate.errorEvents.filterNotNull().asLiveData()
 
-    /** the event subscription */
-    private var eventSubscription: Disposable = EMPTY_DISPOSABLE
+    override val typingUpdates: LiveData<TypingEvent> = delegate.typingUpdates.asLiveData()
 
-    /** stores the mapping from cid to ChannelController */
-    private val activeChannelMapImpl: ConcurrentHashMap<String, ChannelControllerImpl> = ConcurrentHashMap()
-
-    override val typingUpdates: LiveData<TypingEvent> = _typingChannels
-
-    private val activeQueryMapImpl: ConcurrentHashMap<String, QueryChannelsControllerImpl> = ConcurrentHashMap()
-
-    internal val eventHandler: EventHandlerImpl = EventHandlerImpl(this)
-
-    private var logger = ChatLogger.get("Domain")
-    private val cleanTask = object : Runnable {
-        override fun run() {
-            clean()
-            mainHandler.postDelayed(this, 1000)
-        }
-    }
+    internal val eventHandler: EventHandlerImpl
+        get() = delegate.eventHandler
 
     internal lateinit var repos: RepositoryFacade
-    private val syncStateFlow: MutableStateFlow<SyncState?> = MutableStateFlow(null)
     internal var initJob: Deferred<SyncState?>? = null
 
     /** The retry policy for retrying failed requests */
-    override var retryPolicy: RetryPolicy = DefaultRetryPolicy()
-
-    private fun clearState() {
-        _initialized.value = false
-        _online.value = false
-        _totalUnreadCount.value = 0
-        _channelUnreadCount.value = 0
-        _banned.value = false
-        _mutedUsers.value = emptyList()
-        activeChannelMapImpl.clear()
-        activeQueryMapImpl.clear()
-    }
-
-    private fun isTestRunner(): Boolean {
-        return Build.FINGERPRINT.toLowerCase().contains("robolectric")
-    }
-
-    internal fun setUser(user: User) {
-        clearState()
-
-        currentUser = user
-
-        repos = RepositoryFacadeBuilder {
-            context(appContext)
-            database(db)
-            currentUser(currentUser)
-            scope(scope)
-            defaultConfig(defaultConfig)
-            setOfflineEnabled(offlineEnabled)
-        }
-            .build()
-
-        // load channel configs from Room into memory
-        initJob = scope.async {
-            // fetch the configs for channels
-            repos.cacheChannelConfigs()
-
-            // load the current user from the db
-            val syncState = repos.selectSyncState(currentUser.id) ?: SyncState(currentUser.id)
-            // set active channels and recover
-            // restore channels
-            syncState.activeChannelIds.forEach(::channel)
-            // restore queries
-            repos.selectQueriesChannelsByIds(syncState.activeQueryIds)
-                .forEach { spec -> queryChannels(spec.filter, spec.sort) }
-
-            // retrieve the last time the user marked all as read and handle it as an event
-            syncState.markedAllReadAt
-                ?.let { MarkAllReadEvent(user = currentUser, createdAt = it) }
-                ?.let { eventHandler.handleEvent(it) }
-
-            syncState.also { syncStateFlow.value = it }
+    override var retryPolicy: RetryPolicy
+        get() = delegate.retryPolicy
+        set(value) {
+            delegate.retryPolicy = value
         }
 
-        if (client.isSocketConnected()) {
-            setOnline()
-        }
-        startListening()
-        initClean()
-    }
+    internal fun setUser(user: User) = delegate.setUser(user)
 
     internal val job = SupervisorJob()
     internal var scope = CoroutineScope(job + DispatcherProvider.IO)
 
-    init {
-        logger.logI("Initializing ChatDomain with version " + getVersion())
-
-        // if the user is already defined, just call setUser ourselves
-        val current = userOverwrite ?: client.getCurrentUser()
-        if (current != null) {
-            setUser(current)
-        }
-        // past behaviour was to set the user on the chat domain
-        // the new syntax is to automatically pick up changes from the client
-        if (userOverwrite == null) {
-            // listen to future user changes
-            client.preSetUserListeners.add {
-                setUser(it)
-            }
-            // disconnect if the low level client disconnects
-            client.disconnectListeners.add {
-                scope.launch {
-                    disconnect()
-                }
-            }
-        }
-        storeBgSyncDataWhenUserConnects()
-    }
-
     internal suspend fun updateCurrentUser(me: User) {
-        if (me.id != currentUser.id) {
-            throw InputMismatchException("received connect event for user with id ${me.id} while chat domain is configured for user with id ${currentUser.id}. create a new chatdomain when connecting a different user.")
-        }
-        currentUser = me
-        repos.insertCurrentUser(me)
-        _mutedUsers.value = me.mutes
-        setTotalUnreadCount(me.totalUnreadCount)
-        setChannelUnreadCount(me.unreadChannels)
-
-        setBanned(me.banned)
+        delegate.updateCurrentUser(me)
     }
 
     internal suspend fun storeSyncState(): SyncState? {
-        syncStateFlow.value?.let { _syncState ->
-            val newSyncState = _syncState.copy(
-                activeChannelIds = activeChannelMapImpl.keys().toList(),
-                activeQueryIds = activeQueryMapImpl.values.map { it.queryChannelsSpec.id }
-            )
-            repos.insertSyncState(newSyncState)
-            syncStateFlow.value = newSyncState
-        }
-
-        return syncStateFlow.value
+        return delegate.storeSyncState()
     }
 
-    override suspend fun disconnect() {
-        storeSyncState()
-        job.cancelChildren()
-        stopListening()
-        stopClean()
-    }
+    override suspend fun disconnect() = delegate.disconnect()
 
-    override fun getVersion(): String {
-        return STREAM_CHAT_VERSION + "-" + BuildConfig.BUILD_TYPE
-    }
+    override fun getVersion(): String = delegate.getVersion()
 
-    private fun stopClean() {
-        mainHandler.removeCallbacks(cleanTask)
-    }
+    suspend fun <T : Any> runAndRetry(runnable: () -> Call<T>): Result<T> = delegate.runAndRetry(runnable)
 
-    private fun initClean() {
-        mainHandler.postDelayed(cleanTask, 5000)
-    }
+    suspend fun createChannel(c: Channel): Result<Channel> = delegate.createChannel(c)
 
-    suspend fun <T : Any> runAndRetry(runnable: () -> Call<T>): Result<T> {
-        var attempt = 1
-        var result: Result<T>
+    fun addError(error: ChatError) = delegate.addError(error)
 
-        while (true) {
-            result = runnable().execute()
-            if (result.isSuccess || result.error().isPermanent()) {
-                break
-            } else {
-                // retry logic
-                val shouldRetry = retryPolicy.shouldRetry(client, attempt, result.error())
-                val timeout = retryPolicy.retryTimeout(client, attempt, result.error())
+    fun isActiveChannel(cid: String): Boolean = delegate.isActiveChannel(cid)
 
-                if (shouldRetry) {
-                    // temporary failure, continue
-                    logger.logI("API call failed (attempt $attempt), retrying in $timeout seconds. Error was ${result.error()}")
-                    delay(timeout.toLong())
-                    attempt += 1
-                } else {
-                    logger.logI("API call failed (attempt $attempt). Giving up for now, will retry when connection recovers. Error was ${result.error()}")
-                    break
-                }
-            }
-        }
-        // permanent failure case return
-        return result
-    }
+    fun setChannelUnreadCount(newCount: Int) = delegate.setChannelUnreadCount(newCount)
 
-    suspend fun createChannel(c: Channel): Result<Channel> =
-        try {
-            val online = isOnline()
-            c.createdAt = c.createdAt ?: Date()
-            c.syncStatus = if (online) {
-                SyncStatus.IN_PROGRESS
-            } else {
-                SyncStatus.SYNC_NEEDED
-            }
-            if (c.createdBy != currentUser) {
-                c.createdBy = currentUser
-            }
+    fun setBanned(newBanned: Boolean) = delegate.setBanned(newBanned)
 
-            // update livedata
-            val channelController = channel(c.cid)
-            channelController.updateLiveDataFromChannel(c)
+    fun setTotalUnreadCount(newCount: Int) = delegate.setTotalUnreadCount(newCount)
 
-            // Update Room State
-            repos.insertChannel(c)
-
-            // Add to query controllers
-            for (query in activeQueryMapImpl.values) {
-                query.addChannelIfFilterMatches(c)
-            }
-
-            // make the API call and follow retry policy
-            if (online) {
-                val runnable = {
-                    val members = c.members.map { it.getUserId() }
-                    client.createChannel(c.type, c.id, members, c.extraData)
-                }
-                val result = runAndRetry(runnable)
-                if (result.isSuccess) {
-                    c.syncStatus = SyncStatus.COMPLETED
-                    repos.insertChannel(c)
-                    Result(result.data())
-                } else {
-                    if (result.error().isPermanent()) {
-                        c.syncStatus = SyncStatus.FAILED_PERMANENTLY
-                    } else {
-                        c.syncStatus = SyncStatus.SYNC_NEEDED
-                    }
-                    repos.insertChannel(c)
-                    Result(result.error())
-                }
-            } else {
-                Result(c)
-            }
-        } catch (e: IllegalStateException) {
-            Result(ChatError(cause = e))
-        }
-
-    fun addError(error: ChatError) {
-        _errorEvent.value = Event(error)
-    }
-
-    fun isActiveChannel(cid: String): Boolean {
-        return activeChannelMapImpl.containsKey(cid)
-    }
-
-    fun setChannelUnreadCount(newCount: Int) {
-        _channelUnreadCount.value = newCount
-    }
-
-    fun setBanned(newBanned: Boolean) {
-        _banned.value = newBanned
-    }
-
-    fun setTotalUnreadCount(newCount: Int) {
-        _totalUnreadCount.value = newCount
-    }
-
-    private fun storeBgSyncDataWhenUserConnects() {
-        client.addSocketListener(
-            object : SocketListener() {
-                override fun onConnected(event: ConnectedEvent) {
-                    storeBgSyncData()
-                    client.removeSocketListener(this)
-                }
-            }
-        )
-    }
-
-    private fun storeBgSyncData() {
-        if (backgroundSyncEnabled && !isTestRunner()) {
-            val config = BackgroundSyncConfig(client.config.apiKey, currentUser.id, client.getCurrentToken() ?: "")
-            if (config.isValid()) {
-                syncModule.encryptedBackgroundSyncConfigStore.apply {
-                    put(config)
-                }
-            }
-        }
-    }
-
-    /**
-     * Start listening to chat events and keep the room database in sync
-     */
-    private fun startListening() {
-        if (eventSubscription.isDisposed) {
-            eventSubscription = client.subscribe {
-                eventHandler.handleEvents(listOf(it))
-            }
-        }
-    }
-
-    /**
-     * Stop listening to chat events
-     */
-    private fun stopListening() {
-        eventSubscription.dispose()
-    }
-
-    internal fun channel(c: Channel): ChannelControllerImpl {
-        return channel(c.type, c.id)
-    }
+    internal fun channel(c: Channel): ChannelControllerImpl = channel(c.type, c.id)
 
     internal fun channel(cid: String): ChannelControllerImpl {
         if (!CHANNEL_CID_REGEX.matches(cid)) {
@@ -505,56 +194,29 @@ internal class ChatDomainImpl internal constructor(
         channelType: String,
         channelId: String,
     ): ChannelControllerImpl {
-        val cid = "%s:%s".format(channelType, channelId)
-        if (!activeChannelMapImpl.containsKey(cid)) {
-            val channelController =
-                ChannelControllerImpl(
-                    channelType,
-                    channelId,
-                    client,
-                    this
-                )
-            activeChannelMapImpl[cid] = channelController
-            scope.launch(DispatcherProvider.Main) {
-                addTypingChannel(channelController)
-            }
-        }
-        return activeChannelMapImpl.getValue(cid)
+        return ChannelControllerImpl(delegate.channel(channelType, channelId))
     }
 
-    internal fun allActiveChannels(): List<ChannelControllerImpl> =
-        activeChannelMapImpl.values.toList()
-
-    fun generateMessageId(): String {
-        return currentUser.id + "-" + UUID.randomUUID().toString()
+    internal fun allActiveChannels(): List<ChannelControllerImpl> {
+        return delegate.allActiveChannels().map { ChannelControllerImpl(it) }
     }
 
-    private fun addTypingChannel(channelController: ChannelControllerImpl) {
-        _typingChannels.addSource(channelController.typing, _typingChannels::postValue)
-    }
+    fun generateMessageId(): String = delegate.generateMessageId()
 
-    internal fun setOffline() {
-        _online.value = false
-    }
+    internal fun setOffline() = delegate.setOffline()
 
-    internal fun setOnline() {
-        _online.value = true
-    }
+    internal fun setOnline() = delegate.setOnline()
 
-    internal fun setInitialized() {
-        _initialized.value = true
-    }
+    internal fun setInitialized() = delegate.setInitialized()
 
-    override fun isOnline(): Boolean = _online.value
+    override fun isOnline(): Boolean = delegate.isOnline()
 
-    override fun isOffline(): Boolean = !_online.value
+    override fun isOffline(): Boolean = delegate.isOffline()
 
-    override fun isInitialized(): Boolean {
-        return _initialized.value
-    }
+    override fun isInitialized(): Boolean = delegate.isInitialized()
 
     override fun getActiveQueries(): List<QueryChannelsControllerImpl> {
-        return activeQueryMapImpl.values.toList()
+        return delegate.getActiveQueries().map { QueryChannelsControllerImpl(it) }
     }
 
     /**
@@ -565,18 +227,9 @@ internal class ChatDomainImpl internal constructor(
     fun queryChannels(
         filter: FilterObject,
         sort: QuerySort<Channel>,
-    ): QueryChannelsControllerImpl =
-        activeQueryMapImpl.getOrPut("${filter.hashCode()}-${sort.hashCode()}") {
-            QueryChannelsControllerImpl(
-                filter,
-                sort,
-                client,
-                this
-            )
-        }
-
-    private fun queryEvents(cids: List<String>): Result<List<ChatEvent>> =
-        client.getSyncHistory(cids, syncStateFlow.value?.lastSyncedAt ?: Date()).execute()
+    ): QueryChannelsControllerImpl {
+        return QueryChannelsControllerImpl(delegate.queryChannels(filter, sort))
+    }
 
     /**
      * replay events for all active channels
@@ -585,31 +238,11 @@ internal class ChatDomainImpl internal constructor(
      * @param cid ensures that the channel with this id is active
      */
     suspend fun replayEventsForActiveChannels(cid: String? = null): Result<List<ChatEvent>> {
-        // wait for the active channel info to load
-        initJob?.join()
-        // make a list of all channel ids
-        val cids = activeChannelMapImpl.keys().toList().toMutableList()
-        cid?.let {
-            channel(it)
-            cids.add(it)
-        }
-
-        return replayEventsForChannels(cids)
+        return delegate.replayEventsForActiveChannels(cid)
     }
 
     internal suspend fun replayEventsForChannels(cids: List<String>): Result<List<ChatEvent>> {
-        val now = Date()
-
-        return if (cids.isNotEmpty()) {
-            queryEvents(cids).also { resultChatEvent ->
-                if (resultChatEvent.isSuccess) {
-                    eventHandler.updateOfflineStorageFromEvents(resultChatEvent.data())
-                    syncStateFlow.value?.let { syncStateFlow.value = it.copy(lastSyncedAt = now) }
-                }
-            }
-        } else {
-            Result(emptyList())
-        }
+        return delegate.replayEventsForChannels(cids)
     }
 
     /**
@@ -626,224 +259,47 @@ internal class ChatDomainImpl internal constructor(
      * - event recovery for those channels
      * - API calls to create local channels, messages and reactions
      */
-    suspend fun connectionRecovered(recoverAll: Boolean = false) {
-        // 0 ensure load is complete
-        initJob?.join()
+    suspend fun connectionRecovered(recoverAll: Boolean = false) = delegate.connectionRecovered(recoverAll)
 
-        // 1 update the results for queries that are actively being shown right now
-        val updatedChannelIds = mutableSetOf<String>()
-        val queriesToRetry = activeQueryMapImpl.values
-            .toList()
-            .filter { it.recoveryNeeded || recoverAll }
-            .take(3)
-        for (queryChannelController in queriesToRetry) {
-            val pagination = QueryChannelsPaginationRequest(
-                QuerySort<Channel>(),
-                INITIAL_CHANNEL_OFFSET,
-                CHANNEL_LIMIT,
-                MESSAGE_LIMIT,
-                MEMBER_LIMIT
-            )
-            val response = queryChannelController.runQueryOnline(pagination)
-            if (response.isSuccess) {
-                queryChannelController.updateChannelsAndQueryResults(response.data(), pagination.isFirstPage)
-                updatedChannelIds.addAll(response.data().map { it.cid })
-            }
-        }
-        // 2 update the data for all channels that are being show right now...
-        // exclude ones we just updated
-        val cids: List<String> = activeChannelMapImpl
-            .entries
-            .asSequence()
-            .filter { it.value.recoveryNeeded || recoverAll }
-            .filterNot { updatedChannelIds.contains(it.key) }
-            .take(30)
-            .map { it.key }
-            .toList()
-
-        val online = isOnline()
-        logger.logI("recovery called: recoverAll: $recoverAll, online: $online retrying ${queriesToRetry.size} queries and ${cids.size} channels")
-
-        var missingChannelIds = listOf<String>()
-        if (cids.isNotEmpty() && online) {
-            val filter = `in`("cid", cids)
-            val request = QueryChannelsRequest(filter, 0, 30)
-            val result = client.queryChannels(request).execute()
-            if (result.isSuccess) {
-                val channels = result.data()
-                val foundChannelIds = channels.map { it.id }
-                for (c in channels) {
-                    val channelController = this.channel(c)
-                    channelController.updateLiveDataFromChannel(c)
-                }
-                missingChannelIds = cids.filterNot { foundChannelIds.contains(it) }
-                storeStateForChannels(channels)
-            }
-            // create channels that are not present on the API
-            for (c in missingChannelIds) {
-                val channelController = this.channel(c)
-                channelController.watch()
-            }
-            // 3 recover events
-            replayEventsForChannels(cids)
-        }
-
-        // 4 retry any failed requests
-        if (online) {
-            retryFailedEntities()
-        }
-    }
-
-    internal suspend fun retryFailedEntities() {
-        delay(1000)
-        // retry channels, messages and reactions in that order..
-        val channels = retryChannels()
-        val messages = retryMessages()
-        val reactions = retryReactions()
-        logger.logI("Retried ${channels.size} channel entities, ${messages.size} messages and ${reactions.size} reaction entities")
-    }
+    internal suspend fun retryFailedEntities() = delegate.retryFailedEntities()
 
     @VisibleForTesting
-    internal suspend fun retryChannels(): List<Channel> {
-        return repos.selectChannelsSyncNeeded().onEach { channel ->
-            val result = client.createChannel(
-                channel.type,
-                channel.id,
-                channel.members.map(UserEntity::getUserId),
-                channel.extraData
-            ).await()
-
-            when {
-                result.isSuccess -> {
-                    channel.syncStatus = SyncStatus.COMPLETED
-                    repos.insertChannel(channel)
-                }
-                result.isError && result.error().isPermanent() -> {
-                    channel.syncStatus = SyncStatus.FAILED_PERMANENTLY
-                    repos.insertChannel(channel)
-                }
-            }
-        }
-    }
+    internal suspend fun retryChannels(): List<Channel> = delegate.retryChannels()
 
     @VisibleForTesting
-    internal suspend fun retryMessages(): List<Message> {
-        val messages = repos.selectMessagesSyncNeeded()
-        for (message in messages) {
-            val channelClient = client.channel(message.cid)
-            // support sending, deleting and editing messages here
-            val result = when {
-                message.deletedAt != null -> channelClient.deleteMessage(message.id).execute()
-                message.updatedAt != null || message.updatedLocallyAt != null -> {
-                    client.updateMessage(message).execute()
-                }
-                else -> channelClient.sendMessage(message).execute()
-            }
-
-            if (result.isSuccess) {
-                // TODO: 1.1 image upload support
-                repos.insertMessage(message.copy(syncStatus = SyncStatus.COMPLETED))
-            } else if (result.isError && result.error().isPermanent()) {
-                repos.insertMessage(message.copy(syncStatus = SyncStatus.FAILED_PERMANENTLY))
-            }
-        }
-
-        return messages
-    }
+    internal suspend fun retryMessages(): List<Message> = delegate.retryMessages()
 
     @VisibleForTesting
-    internal suspend fun retryReactions(): List<Reaction> {
-        return repos.selectReactionsSyncNeeded().onEach { reaction ->
-            val result = if (reaction.deletedAt != null) {
-                client.deleteReaction(reaction.messageId, reaction.type).execute()
-            } else {
-                client.sendReaction(reaction, reaction.enforceUnique).execute()
-            }
+    internal suspend fun retryReactions(): List<Reaction> = delegate.retryReactions()
 
-            if (result.isSuccess) {
-                reaction.syncStatus = SyncStatus.COMPLETED
-                repos.insertReaction(reaction)
-            } else if (result.error().isPermanent()) {
-                reaction.syncStatus = SyncStatus.FAILED_PERMANENTLY
-                repos.insertReaction(reaction)
-            }
-        }
-    }
-
-    suspend fun storeStateForChannel(channel: Channel) {
-        return storeStateForChannels(listOf(channel))
-    }
+    suspend fun storeStateForChannel(channel: Channel) = delegate.storeStateForChannel(channel)
 
     suspend fun storeStateForChannels(channelsResponse: Collection<Channel>) {
-        val users = mutableMapOf<String, User>()
-        val configs: MutableCollection<ChannelConfig> = mutableSetOf()
-        // start by gathering all the users
-        val messages = mutableListOf<Message>()
-        for (channel in channelsResponse) {
-
-            users.putAll(channel.users().associateBy { it.id })
-            configs += ChannelConfig(channel.type, channel.config)
-
-            channel.messages.forEach { message ->
-                message.enrichWithCid(channel.cid)
-                users.putAll(message.users().associateBy { it.id })
-            }
-
-            messages.addAll(channel.messages)
-        }
-
-        repos.storeStateForChannels(
-            configs = configs,
-            users = users.values.toList(),
-            channels = channelsResponse,
-            messages = messages
-        )
-
-        logger.logI("storeStateForChannels stored ${channelsResponse.size} channels, ${configs.size} configs, ${users.size} users and ${messages.size} messages")
+        return delegate.storeStateForChannels(channelsResponse)
     }
 
     suspend fun selectAndEnrichChannel(
         channelId: String,
         pagination: QueryChannelPaginationRequest,
     ): Channel? {
-        return selectAndEnrichChannels(listOf(channelId), pagination.toAnyChannelPaginationRequest()).getOrNull(0)
+        return delegate.selectAndEnrichChannel(channelId, pagination)
     }
 
     suspend fun selectAndEnrichChannel(
         channelId: String,
         pagination: QueryChannelsPaginationRequest,
     ): Channel? {
-        return selectAndEnrichChannels(listOf(channelId), pagination.toAnyChannelPaginationRequest()).getOrNull(0)
+        return delegate.selectAndEnrichChannel(channelId, pagination)
     }
 
     suspend fun selectAndEnrichChannels(
         channelIds: List<String>,
         pagination: QueryChannelsPaginationRequest,
     ): List<Channel> {
-        return selectAndEnrichChannels(channelIds, pagination.toAnyChannelPaginationRequest())
+        return delegate.selectAndEnrichChannels(channelIds, pagination)
     }
 
-    private suspend fun selectAndEnrichChannels(
-        channelIds: List<String>,
-        pagination: AnyChannelPaginationRequest,
-    ): List<Channel> {
-        return repos.selectChannels(channelIds, pagination).applyPagination(pagination)
-    }
+    override fun clean() = delegate.clean()
 
-    override fun clean() {
-        for (channelController in activeChannelMapImpl.values.toList()) {
-            channelController.clean()
-        }
-    }
-
-    override fun getChannelConfig(channelType: String): Config {
-        return repos.selectChannelConfig(channelType)?.config ?: defaultConfig
-    }
-
-    companion object {
-        val EMPTY_DISPOSABLE = object : Disposable {
-            override val isDisposed: Boolean = true
-            override fun dispose() {}
-        }
-    }
+    override fun getChannelConfig(channelType: String): Config = delegate.getChannelConfig(channelType)
 }
