+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/usecase/CreateChannelImpl.kt
@@ -1,0 +1,26 @@
+package io.getstream.chat.android.offline.usecase
+
+import androidx.annotation.CheckResult
+import io.getstream.chat.android.client.call.Call
+import io.getstream.chat.android.client.call.CoroutineCall
+import io.getstream.chat.android.client.models.Channel
+import io.getstream.chat.android.offline.ChatDomainImpl as NewChatDomainImpl
+
+public interface CreateChannel {
+    /**
+     * Creates a new channel. Will retry according to the retry policy if it fails
+     * @see io.getstream.chat.android.livedata.utils.RetryPolicy
+     *
+     * @param channel the channel object
+     */
+    @CheckResult
+    public operator fun invoke(channel: Channel): Call<Channel>
+}
+
+internal class CreateChannelImpl(private val domainImpl: NewChatDomainImpl) : CreateChannel {
+    override operator fun invoke(channel: Channel): Call<Channel> {
+        return CoroutineCall(domainImpl.scope) {
+            domainImpl.createChannel(channel)
+        }
+    }
+}
