+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/usecase/KeystrokeImpl.kt
@@ -1,0 +1,32 @@
+package io.getstream.chat.android.offline.usecase
+
+import androidx.annotation.CheckResult
+import io.getstream.chat.android.client.call.Call
+import io.getstream.chat.android.client.call.CoroutineCall
+import io.getstream.chat.android.livedata.utils.validateCid
+import io.getstream.chat.android.offline.ChatDomainImpl as NewChatDomainImpl
+
+public interface Keystroke {
+    /**
+     * Keystroke should be called whenever a user enters text into the message input
+     * It automatically calls stopTyping when the user stops typing after 5 seconds
+     *
+     * @param cid the full channel id i. e. messaging:123
+     * @param parentId set this field to `message.id` to indicate that typing event is happening in a thread
+     *
+     * @return True when a typing event was sent, false if it wasn't sent
+     */
+    @CheckResult
+    public operator fun invoke(cid: String, parentId: String? = null): Call<Boolean>
+}
+
+internal class KeystrokeImpl(private val domainImpl: NewChatDomainImpl) : Keystroke {
+    override operator fun invoke(cid: String, parentId: String?): Call<Boolean> {
+        validateCid(cid)
+
+        val channelController = domainImpl.channel(cid)
+        return CoroutineCall(domainImpl.scope) {
+            channelController.keystroke(parentId)
+        }
+    }
+}
