+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/usecase/SendGiphyImpl.kt
@@ -1,0 +1,32 @@
+package io.getstream.chat.android.offline.usecase
+
+import androidx.annotation.CheckResult
+import io.getstream.chat.android.client.call.Call
+import io.getstream.chat.android.client.call.CoroutineCall
+import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.livedata.utils.validateCid
+import io.getstream.chat.android.offline.ChatDomainImpl as NewChatDomainImpl
+
+public interface SendGiphy {
+    /**
+     * Sends selected giphy message to the channel. Removes the original "ephemeral" message from local storage.
+     * Returns new "ephemeral" message with new giphy url.
+     * API call to remove the message is retried according to the retry policy specified on the chatDomain
+     * @param message the message to send
+     * @see io.getstream.chat.android.livedata.utils.RetryPolicy
+     */
+    @CheckResult
+    public operator fun invoke(message: Message): Call<Message>
+}
+
+internal class SendGiphyImpl(private val domainImpl: NewChatDomainImpl) : SendGiphy {
+    override operator fun invoke(message: Message): Call<Message> {
+        val cid = message.cid
+        validateCid(cid)
+
+        val channelController = domainImpl.channel(cid)
+        return CoroutineCall(domainImpl.scope) {
+            channelController.sendGiphy(message)
+        }
+    }
+}
