+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/usecase/SetMessageForReply.kt
@@ -1,0 +1,32 @@
+package io.getstream.chat.android.offline.usecase
+
+import androidx.annotation.CheckResult
+import io.getstream.chat.android.client.call.Call
+import io.getstream.chat.android.client.call.CoroutineCall
+import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.client.utils.Result
+import io.getstream.chat.android.livedata.utils.validateCid
+import io.getstream.chat.android.offline.ChatDomainImpl as NewChatDomainImpl
+
+public interface SetMessageForReply {
+    /**
+     * Set the reply state for the channel.
+     *
+     * @param cid CID of the channel where reply state is being set.
+     * @param message The message we want reply to. The null value means dismiss reply state.
+     */
+    @CheckResult
+    public operator fun invoke(cid: String, message: Message?): Call<Unit>
+}
+
+internal class SetMessageForReplyImpl(private val chatDomain: NewChatDomainImpl) : SetMessageForReply {
+    override fun invoke(cid: String, message: Message?): Call<Unit> {
+        validateCid(cid)
+
+        val channelController = chatDomain.channel(cid)
+        return CoroutineCall(chatDomain.scope) {
+            channelController.replyMessage(message)
+            Result(Unit)
+        }
+    }
+}
