+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/usecase/ThreadLoadMoreImpl.kt
@@ -1,0 +1,34 @@
+package io.getstream.chat.android.offline.usecase
+
+import androidx.annotation.CheckResult
+import io.getstream.chat.android.client.call.Call
+import io.getstream.chat.android.client.call.CoroutineCall
+import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.livedata.utils.validateCid
+import io.getstream.chat.android.offline.ChatDomainImpl as NewChatDomainImpl
+
+public interface ThreadLoadMore {
+    /**
+     * Loads more messages for the specified thread
+     *
+     * @param cid: the full channel id i. e. messaging:123
+     * @param parentId: the parentId of the thread
+     * @param messageLimit: how many new messages to load
+     */
+    @CheckResult
+    public operator fun invoke(cid: String, parentId: String, messageLimit: Int): Call<List<Message>>
+}
+
+internal class ThreadLoadMoreImpl(private val domainImpl: NewChatDomainImpl) : ThreadLoadMore {
+    override operator fun invoke(cid: String, parentId: String, messageLimit: Int): Call<List<Message>> {
+        validateCid(cid)
+        require(parentId.isNotEmpty()) { "parentId can't be empty" }
+
+        val channelController = domainImpl.channel(cid)
+        val threadController = channelController.getThread(parentId)
+
+        return CoroutineCall(domainImpl.scope) {
+            threadController.loadOlderMessages(messageLimit)
+        }
+    }
+}
